Cast Time objects to integers before comparison
There's a recurring gotcha here where the Time object stores time to the nanosecond, but the database does not.

Therefore, when we retreive a timestamp back from a persisted object and compare it against our Time object, they are not equal.

We can get around this by comparing the Time objects as integers.