Refactor multi-step form components
This extracts out the multi-step form logic and behaviour from the BuyerApplication update method into a new set of concerns which can be shared across different kinds of operations.

I've created two DSLs here for ease of use in defining the specifics for a multi-step form, like the I18n key and models to use.