Remove text mailer layout as unnecessary
We're using the rails premailer gem which by default converts html
emails to plain text emails. It does a pretty good job so it would
be sensible to work with this and only design/write html emails in the
short term as it will allow us to move faster (not having to make
the same content change in two different places).