Convert form objects into contracts
This converts the original seller application form objects to become Trailblazer contracts.

There are few actual changes which need to be made to the objects, as Trailblazer contracts still inherit from `Reform::Form`.

The `Base` class now introduces some of the included modules in line with the equivalent class in BuyerOnboarding. It may be that these need to be tweaked further down the line once we get to fully testing the refactored flow.

This also renames and updates the unit specs so that they continue to work.