Refactor away some of the Trailblazer overrides
This prevents Trailblazer automatically wrapping the form object in its `Trailblazer::Rails::Form` class – which otherwise would break the i18n lookups.

Instead, the form object is passed in the controller directly from the Operation into the view.

This means we no longer need to monkey patch the form object in the initializer.