Fix error handling for nested fields
Trailblazer operations appear to handle differently the storage of error messages on nested objects.

This is mostly a good thing, as it validates the objects with less hackery than before.

However (there's always a catch with TRB) – this breaks the error handling here as Reform no longer likes receiving errors as hashes.

This commit changes the error handling to store a single generic error message against the field name when a nested child has an error.

(This appears to fix the problem of addresses not showing a message when you don't start to add one.)