Redesign the seller onboarding journey
This redesigns the seller onboarding journey to support multiple question sets and a progress dashboard.

There's still a lot more to do on this, but merging now to avoid blocking work on other branches.

The feature spec has been updated, however there are many failing contract specs which are currently being skipped.