Refactor SellerApplication creation/submission
This converts the creation of seller applications into an Operation, as per the creation of buyer applications.

As a result, the Sellers::Applications::RootController has been tidied up, which has resulted in a change to the Submit operation. I've used this as a chance to write more tests for this operation.