First pass at inviting another user
This adds the ability to invite another user to your seller account.

When an invitation is sent, a new user account is created in the data for the invitee. The account requires email confirmation, so we send a custom confirmation email to the user with a link to confirm.

At present, we randomly set a password, so we will need to introduce some custom behaviour to handle setting a password and confirming the account for this specific use case.