Raise exception when pagination called without args
This causes the Search class to raise an exception when a pagination method is invoked without the `page` argument present.

If you don't provide the `page` argument, Kaminari can't provide its helper methods and will cause errors further in the views.

The Search class isn't opinionated about pagination and we don't want to set defaults – that's a responsibility for the controller. If you need to see all the results without pagination, just call the `results` public method instead.