Replace Poltergeist with headless Chrome
This swaps Poltergeist for headless Chrome when running feature specs. Poltergeist was having issues with running JavaScript and it was about time we moved over.

This also updates the feature specs which were using `trigger('click')` as a workaround, to now just call `click` directly. It also fixes an issue which was preventing checkboxes in search filters from being properly clicked by the headless driver.

Fixes #1