Selectively show product profile fields
This changes how product profile fields are displayed to actually check the values of conditional fields.

This avoids a problem where a seller may complete an optional field, but later change a value which causes that field to no longer be displayed. In this scenario, the field would still exist in the record and be visible to the user.