Restrict seller onboarding steps based on status
This changes the seller onboarding flow so that there are actual access controls to each step based on whether the initial questions ("tailor") have been completed or not.

This also restricts access to the 'tailor' questions once the step has been completed.

A new boolean field has been added to seller_applications to keep track of whether the tailor fields have been completed or not.