Convert attachments to separate model
This changes how files are attached to the Seller model. A new model, Document, has been created, which represents a document attachment. This can contain other metadata, like the virus scanning status.

Documents can be added to any kind of model using the 'documentable' concern, which adds helper fields to make forms work properly.

This also updates the forms so that seller onboarding continues to work.