Remove term keyword argument from Search::Seller
We shouldn't always require a term to be provided to the Seller::Search class. This made sense for the frontend search, but not for other uses.

This brings the behaviour in line with search classes we built for the Ops view. Backwards compatibility is maintained with existing implementations by merging the `term` argument into the `selected_filters` list.