Only show an add link to an expanding list when not full
Right now, if the expanding list has 10 items, loading the page for the first time will still include a link to add another item.

This adds a check on the code inserting the link, to skip if the list is already full.

The existing behaviour to insert the link when the list is no longer full will continue to allow users to remove an item and then add another one.