Lay groundwork for renaming seller applications to versions
This changes the SellerApplication model to dynamically set the model's table name, based on whether the 'seller_versions' table exists or not.

It's important we deploy this before adding the migration because, when the migrations are run, it can be ~20 minutes before the code changes to reflect the updated table name have been propagated across all hosts.